backend/requirements.txt
# Flask and Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6

# Firebase
firebase-admin==6.3.0

# Twilio for SMS
twilio==8.10.0

# Web Server - Using threading mode instead of eventlet for Python 3.13 compatibility
python-socketio==5.10.0

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Utilities
pytz==2023.3
